+++ backend/service/base_api/currency/find_currency_all_logic.v
@@ -40,7 +40,7 @@ pub struct CurrencyListReq {
 	page_size     int    @[json: 'pageSize']
 	english_name  string @[json: 'englishName']
 	currency_code string @[json: 'currencyCode']
-	status        []u8   @[json: 'status']
+	status        []i16  @[json: 'status']
 }
 
 pub struct CurrencyData {
@@ -49,12 +49,12 @@ pub struct CurrencyData {
 	simplified_name           string  @[json: 'simplifiedName']
 	currency_code             string  @[json: 'currencyCode']
 	currency_symbol           string  @[json: 'currencySymbol']
-	decimal_place             u8      @[json: 'decimalPlace']
+	decimal_place             i16     @[json: 'decimalPlace']
 	exchange_rate             f64     @[json: 'exchangeRate']
 	exchange_rate_fluctuation f64     @[json: 'exchangeRateFluctuation']
 	exchange_rate_used        f64     @[json: 'exchangeRateUsed']
 	sort                      ?int    @[json: 'sort']
-	status                    u8      @[json: 'status']
+	status                    i16     @[json: 'status']
 	updater_id                ?string @[json: 'updaterId']
 	creator_id                ?string @[json: 'creatorId']
 	created_at                string  @[json: 'createdAt']
