+++ backend/service/base_api/language/create_language_logic.v
@@ -54,8 +54,8 @@ pub struct CreateLanguageReq {
 	three_letter_code        string @[json: 'threeLetterCode']
 	utf8_encoding            string @[json: 'utf8Encoding']
 	sort                     ?int   @[json: 'sort']
-	status                   u8     @[json: 'status']
-	is_basic                 u8     @[json: 'isBasic']
+	status                   i16    @[json: 'status']
+	is_basic                 i16    @[json: 'isBasic']
 }
 
 pub struct CreateLanguageResp {
